refactor(analytics): align guest snapshot frontend naming
