+++ frontend/src/apis/guestTimetableSnapshotAPI.ts
