+++ frontend/tsconfig.app.json
@@ -29,8 +29,5 @@
   },
   "include": [
     "src"
-  ],
-  "exclude": [
-    "src/**/*.test.ts"
   ]
 }
